Try change variable

+++ lab3_1.cpp
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main()
 {
-	unsigned short int score[5][4], max[2] = {}, sum;
+	unsigned int score[5][4], max[2] = {}, sum;
 	for (int i = 0; i < 5; i++)
 	{
 		scanf("%d %d %d %d", &score[i][0], &score[i][1], &score[i][2], &score[i][3]);
